Implement printTableHeader
This defines the fundamental algorithm for how to iterate over the field widths
and determine the amount of padding to print after arbitrary table cell data.

The comments are diagrammatic and not necessarily the most intuitive. The table
diagram in the comments shows areas of the table where there is interpolation.
There is also the presence of ␠ to indicate padding between the window edges of
the terminal emulator and the table.