Remove the unnecessary exception type InvalidSearchMethodException.
This exception is an uncaught exception, but it is only made during programmer
error. Rather than throwing and catching such an exception, we should simply not
make this programming error when using our own methods. We aren't developing a
library, hence we should never throw any sort of invalid argument exception.